[需求]gl渲染后视频帧软编成功

+++ videoeditor-armv7a/src/main/java/com/meitu/library/videoeditor/save/hardmux/HardMuxJni.java
@@ -21,7 +21,7 @@ public class HardMuxJni implements HardMuxListener {
     }
 
     @Override
-    public void onAudioFrame(byte[] data, double pts) {
+    public void onAudioFrame(byte[] data, long pts) {
 
     }
