+++ videoeditor-armv7a/src/main/java/com/meitu/library/videoeditor/save/hardmux/HardMuxListener.java
@@ -8,7 +8,7 @@ public interface HardMuxListener {
 
    void onVideoFrame(byte[] data, double pts) ;
 
-   void onAudioFrame(byte[] data, double pts) ;
+   void onAudioFrame(byte[] data, long pts) ;
 
    void onVideoDone() ;
